@@ -19,12 +19,11 @@ namespace MZPO.Processors
         private readonly SheetsService _service;
         private readonly string SpreadsheetId;
         private readonly BaseRepository<Lead> leadRepo;
-        private readonly BaseRepository<Company> compRepo;
         private readonly BaseRepository<Contact> contRepo;
         private readonly long endDate;
         protected readonly CancellationToken _token;
 
-        public WeeklyReportProcessor(AmoAccount acc, GSheets gSheets, string spreadsheetId, TaskList processQueue, CancellationToken token, long dateTo)
+        public WeeklyReportProcessor(AmoAccount acc, GSheets gSheets, string spreadsheetId, TaskList processQueue, long dateTo, CancellationToken token)
         {
             _acc = acc;
             _processQueue = processQueue;
@@ -32,7 +31,6 @@ namespace MZPO.Processors
             _service = gSheets.GetService();
             SpreadsheetId = spreadsheetId;
             leadRepo = _acc.GetRepo<Lead>();
-            compRepo = _acc.GetRepo<Company>();
             contRepo = _acc.GetRepo<Contact>();
 
             dataRanges = new List<(int, int)>();
@@ -248,8 +246,10 @@ namespace MZPO.Processors
             #endregion
 
             #region Executing request
-            var batchRequest = new BatchUpdateSpreadsheetRequest();
-            batchRequest.Requests = requestContainer;
+            var batchRequest = new BatchUpdateSpreadsheetRequest
+            {
+                Requests = requestContainer
+            };
 
             await _service.Spreadsheets.BatchUpdate(batchRequest, SpreadsheetId).ExecuteAsync();
             #endregion
@@ -258,11 +258,11 @@ namespace MZPO.Processors
         private void PrepareRow(int sheetId, string A, int B, int C, int D, int E, double H, double I, int J, int K, int L, int M)
         {
             #region Prepare data
-            var rows = new List<RowData>();
-
-            rows.Add(new RowData()
+            var rows = new List<RowData>
             {
-                Values = new List<CellData>(){
+                new RowData()
+                {
+                    Values = new List<CellData>(){
                          new CellData(){
                              UserEnteredValue = new ExtendedValue(){ StringValue = A},
                              UserEnteredFormat = columns["A"] },
@@ -303,7 +303,8 @@ namespace MZPO.Processors
                              UserEnteredValue = new ExtendedValue(){ NumberValue = M},
                              UserEnteredFormat = columns["M"] },
                 }
-            });
+                }
+            };
             #endregion
 
             #region Add request
@@ -326,10 +327,10 @@ namespace MZPO.Processors
             int timeOfReference = (int)lead.created_at;
 
             #region Результат звонка
-            if (lead.custom_fields_values is { })
+            if (lead.custom_fields_values is not null)
             {
                 var cf = lead.custom_fields_values.FirstOrDefault(x => x.field_id == 644675);
-                if (cf is { })
+                if (cf is not null)
                 {
                     var cfValue = (string)cf.values[0].value;
                     if (cfValue == "Принят" || cfValue == "Ручная сделка") return 0;
@@ -340,7 +341,7 @@ namespace MZPO.Processors
             var leadEvents = leadRepo.GetEvents(lead.id);
 
             #region Смена ответственного
-            if ((leadEvents is { }) &&
+            if ((leadEvents is not null) &&
                 leadEvents.Where((x => x.type == "entity_responsible_changed")).Any(x => x.value_before[0].responsible_user.id == 2576764))
                 timeOfReference = (int)leadEvents
                     .Where((x => x.type == "entity_responsible_changed"))
@@ -349,7 +350,7 @@ namespace MZPO.Processors
             #endregion
 
             #region Cообщения в чат
-            if (leadEvents is { })
+            if (leadEvents is not null)
                 foreach (var e in leadEvents)
                     if ((e.type == "outgoing_chat_message") || (e.type == "incoming_chat_message"))
                         replyTimestamps.Add((int)e.created_at);
@@ -357,18 +358,18 @@ namespace MZPO.Processors
 
             #region Исходящее письмо
             var notes = leadRepo.GetNotes(lead.id);
-            if (notes is { })
+            if (notes is not null)
                 foreach (var n in notes)
                     if ((n.note_type == "amomail_message") && (n.parameters.income == false))
                         replyTimestamps.Add((int)n.created_at);
             #endregion
 
             #region Звонки
-            if (lead._embedded.contacts is { })
+            if (lead._embedded.contacts is not null)
                 foreach (var c in lead._embedded.contacts)
                 {
                     var contactEvents = contRepo.GetEvents(c.id);
-                    if (contactEvents is { })
+                    if (contactEvents is not null)
                         foreach (var e in contactEvents)
                         {
                             if ((e.type == "outgoing_call") || (e.type == "incoming_call"))
@@ -376,7 +377,7 @@ namespace MZPO.Processors
                                 var callNote = contRepo.GetNoteById(e.value_after[0].note.id);
                                 int duration = 0;
 
-                                if (callNote is { } && callNote.parameters is { } && callNote.parameters.duration > 0)
+                                if (callNote is not null && callNote.parameters is not null && callNote.parameters.duration > 0)
                                     duration = (int)callNote.parameters.duration;
 
                                 int actualCallTime = (int)e.created_at - duration;
@@ -430,7 +431,7 @@ namespace MZPO.Processors
             foreach (var p in pipelines)
             {
                 var leadsOpened = leadRepo.GetByCriteria($"filter[pipeline_id][0]={p}&filter[created_at][from]={dataRange.Item1}&filter[created_at][to]={dataRange.Item2}&filter[responsible_user_id]={manager.Item1}&with=contacts");
-                if (leadsOpened is { })
+                if (leadsOpened is not null)
                     newLeads.AddRange(leadsOpened);
             }
 
@@ -449,7 +450,7 @@ namespace MZPO.Processors
 
             var leadsClosed = leadRepo.GetByCriteria($"filter[pipeline_id][0]=3198184&filter[closed_at][from]={dataRange.Item1}&filter[closed_at][to]={dataRange.Item2}&filter[responsible_user_id]={manager.Item1}");
 
-            if (leadsClosed is { }) allLeads.AddRange(leadsClosed);
+            if (leadsClosed is not null) allLeads.AddRange(leadsClosed);
 
             _processQueue.UpdateTaskName("report_retail", $"WeeklyReport: {manager.Item2}, {dates}, closed leads: {allLeads}");
             #endregion
@@ -477,7 +478,7 @@ namespace MZPO.Processors
 
             var callIdList = new List<int>();
 
-            if (inCalls is { })
+            if (inCalls is not null)
                 foreach (var e in inCalls.Where(x => x.created_by == manager.Item1))
                     callIdList.Add(e.value_after[0].note.id);
 
@@ -485,12 +486,12 @@ namespace MZPO.Processors
             if (callIdList.Any())
                 callNotes.AddRange(contRepo.BulkGetNotesById(callIdList));
 
-            if (callNotes is { } && callNotes.Any())
+            if (callNotes is not null && callNotes.Any())
                 foreach (var n in callNotes)
                 {
                     int duration = -1;
 
-                    if (n is { } && n.parameters is { })
+                    if (n is not null && n.parameters is not null)
                         duration = (int)n.parameters.duration;
 
                     if (duration == 0) missedCallsCount++;
@@ -514,8 +515,10 @@ namespace MZPO.Processors
             #region Updating sheet
             if (requestContainer.Any())
             {
-                var batchRequest = new BatchUpdateSpreadsheetRequest();
-                batchRequest.Requests = requestContainer;
+                var batchRequest = new BatchUpdateSpreadsheetRequest
+                {
+                    Requests = requestContainer
+                };
 
                 await _service.Spreadsheets.BatchUpdate(batchRequest, SpreadsheetId).ExecuteAsync();
             }
@@ -532,13 +535,17 @@ namespace MZPO.Processors
             _processQueue.UpdateTaskName("report_retail", $"WeeklyReport: {dates}, outgoing calls");
 
             outCalls = new List<Event>();
-            outCalls.AddRange(contRepo.GetEventsByCriteria($"filter[type]=outgoing_call&filter[created_at][from]={dataRange.Item1}&filter[created_at][to]={dataRange.Item2}"));
+            var calls = contRepo.GetEventsByCriteria($"filter[type]=outgoing_call&filter[created_at][from]={dataRange.Item1}&filter[created_at][to]={dataRange.Item2}");
+            if (calls is not null)
+                outCalls.AddRange(calls);
             #endregion
 
             #region Входящие вызовы
             _processQueue.UpdateTaskName("report_retail", $"WeeklyReport: {dates}, incoming calls");
             inCalls = new List<Event>();
-            inCalls.AddRange(contRepo.GetEventsByCriteria($"filter[type]=incoming_call&filter[created_at][from]={dataRange.Item1}&filter[created_at][to]={dataRange.Item2}"));
+            calls = contRepo.GetEventsByCriteria($"filter[type]=incoming_call&filter[created_at][from]={dataRange.Item1}&filter[created_at][to]={dataRange.Item2}");
+            if (calls is not null)
+                inCalls.AddRange(calls);
             #endregion
 
             GC.Collect();
@@ -572,8 +579,10 @@ namespace MZPO.Processors
             #region Update sheet
             if (requestContainer.Any())
             {
-                var batchRequest = new BatchUpdateSpreadsheetRequest();
-                batchRequest.Requests = requestContainer;
+                var batchRequest = new BatchUpdateSpreadsheetRequest
+                {
+                    Requests = requestContainer
+                };
 
                 await _service.Spreadsheets.BatchUpdate(batchRequest, SpreadsheetId).ExecuteAsync();
             }
@@ -587,11 +596,11 @@ namespace MZPO.Processors
             foreach (var m in managers)
             {
                 #region Prepare data
-                var rows = new List<RowData>();
-
-                rows.Add(new RowData()
+                var rows = new List<RowData>
                 {
-                    Values = new List<CellData>(){
+                    new RowData()
+                    {
+                        Values = new List<CellData>(){
                          new CellData(){
                              UserEnteredValue = new ExtendedValue(){ StringValue = $"{m.Item2}"},
                              UserEnteredFormat = columns["A"] },
@@ -632,7 +641,8 @@ namespace MZPO.Processors
                              UserEnteredValue = new ExtendedValue(){ FormulaValue = $"='{m.Item2}'!M{dataRanges.Count + 2}" },
                              UserEnteredFormat = columns["M"] },
                 }
-                });
+                    }
+                };
                 #endregion
 
                 #region Add request
@@ -670,8 +680,10 @@ namespace MZPO.Processors
             #region Update sheet
             if (requestContainer.Any())
             {
-                var batchRequest = new BatchUpdateSpreadsheetRequest();
-                batchRequest.Requests = requestContainer;
+                var batchRequest = new BatchUpdateSpreadsheetRequest
+                {
+                    Requests = requestContainer
+                };
 
                 await _service.Spreadsheets.BatchUpdate(batchRequest, SpreadsheetId).ExecuteAsync();
             }
@@ -723,7 +735,7 @@ namespace MZPO.Processors
                         {
                             SheetId = m.Item1,
                             StartRowIndex = dataRanges.Count + 2,
-                            EndRowIndex = dataRanges.Count + 2 + rows.Count(),
+                            EndRowIndex = dataRanges.Count + 2 + rows.Count,
                             StartColumnIndex = 0,
                             EndColumnIndex = 2
                         }
@@ -735,8 +747,10 @@ namespace MZPO.Processors
             #region Update sheet
             if (requestContainer.Any())
             {
-                var batchRequest = new BatchUpdateSpreadsheetRequest();
-                batchRequest.Requests = requestContainer;
+                var batchRequest = new BatchUpdateSpreadsheetRequest
+                {
+                    Requests = requestContainer
+                };
 
                 await _service.Spreadsheets.BatchUpdate(batchRequest, SpreadsheetId).ExecuteAsync();
             }
